+++ lib/mixpanel.rb
@@ -1,2 +1,8 @@
-require 'mixpanel/mixpanel'
-require 'mixpanel/mixpanel_middleware'
+require 'mixpanel/tracker'
+
+module Mixpanel
+  def self.new(token, env, async = false)
+    Kernel.warn("DEPRECATED: Use Mixpanel::Tracker.new instead")
+    Mixpanel::Tracker.new(token, env, async)
+  end
+end